requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pymupdf==1.25.3
openpyxl==3.1.5
google-genai==1.10.0
pydantic==2.10.4
